Complete q_insert_head()
This commit tried to complete code in function
q_insert_head().

The function will now check if queue is NULL first,
then try to allocate memory for new node. After that,
try to allocate memory for value element in new node,
the size of the memory for value is decided using the
length of the input string.

After all of that, the next element in new head is pointed
to the head of current queue, and the head of the queue is pointed
to the new head. Finally, increase the value of length in queue by 1.

After each call to malloc, the function would check if malloc failed,
and if failed, the function would return false. Otherwise, the function
will continue.