Complete q_size()
We added a new element called `len` in struct queue_t before,
we use this element to track the length of a queue.

If we want to know the length of a queue, instead of going
through the whole queue, we simply just return the value of `len`
in the queue.