Complete q_insert_tail()
Added code to try to complete function q_insert_tail().

The function behaves like q_insert_head(), but instead of
touching the head of a queue, this touches the tail of a queue.

And by using the q->tail, we can get the tail of a queue easily,
don't need to go through the whole queue.