Complete q_remove_head()
This commit tried to complete the function
q_remove_head().

The function will check if the queue is valid first,
then if sp isn't NULL, the function will copy the value
of the node that's going to be deleted to the sp.

After that, the function will detach the node from the queue,
then free the memory being used by the node.

Lastly, decrease the len of the queue by one.