Consider insertion when the queue is empty
This commit fixed issue when inserting node into newly
created queue.

Whether inserting node to head or tail of a queue, if the queue
is newly created, this means either of them will still be NULL
even after insertion. This commit tried to fix this issue.

If we are inserting head into a new queue, then we also assign the
tail of the queue to same node. We do the same thing for inserting
tail, but we assign the head instead of tail.