Fix copying value when removing head
This commit fixed the issue when trying to remove head
node from queue.

The `strncpy` part of code copied one more character than
desired. And the `sp[bufsize]` line also assigned the value
to the memory area which is outside sp.

This commit changed that part of the code, now `strncpy` only copy
`bufsize - 2` amount of characters. And now the null terminator is
assigned to `sp[bufsize-1]`.