Fix issue in q_remove_head()
This commit fixes issue in function q_remove_head().

The `q->head = q->head->next;` line was before we compare
if the element is the last element in the queue. When there's
two elements in the queue, this will turn `q->tail` into `NULL`,
even when there's still node in the queue.