Fix reference error in q_insert_tail()
This commit fixes issue in function q_insert_tail().

When we add a new node to a empty queue, the queue's next should be
NULL before we modify it, and by calling `q->tail->next` at this state,
it will occur an reference error.

To avoid this error, we check if `q->tail` is NULL, and we only call
`q->tail->next` if the value isn't NULL.