Implement q_sort() with bubble sort
This commit added code that will implement bubble sort
in q_sort().

I used three variables `prev`, `curr` and `next` to control
the flow of sort. If the string in `next` node has a larger index
value than the string in `curr` node, then a swap will occur. The
variables will move to next node otherwise.

Although this sort works, running tests showed that this method, or
the code, is too slow. Need to write shorter code, or even change
the sort method.